refac(branch delete): extract handler
