+++ main.go
@@ -21,6 +21,7 @@ import (
 	"go.abhg.dev/gs/internal/forge/gitlab"
 	"go.abhg.dev/gs/internal/git"
 	"go.abhg.dev/gs/internal/handler/checkout"
+	"go.abhg.dev/gs/internal/handler/delete"
 	"go.abhg.dev/gs/internal/handler/submit"
 	"go.abhg.dev/gs/internal/handler/track"
 	"go.abhg.dev/gs/internal/secret"
@@ -387,6 +388,21 @@ func (cmd *mainCmd) AfterApply(ctx context.Context, kctx *kong.Context, logger *
 				OpenRemoteRepository: func(ctx context.Context, remote string) (forge.Repository, error) {
 					return openRemoteRepository(ctx, log, secretStash, forges, wt.Repository(), remote)
 				},
+			}, nil
+		}),
+		kctx.BindSingletonProvider(func(
+			log *silog.Logger,
+			store *state.Store,
+			wt *git.Worktree,
+			svc *spice.Service,
+		) (DeleteHandler, error) {
+			return &delete.Handler{
+				Log:        log,
+				View:       view,
+				Repository: wt.Repository(),
+				Worktree:   wt,
+				Store:      store,
+				Service:    svc,
 			}, nil
 		}),
 	)
