+++ stack_delete.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 
 	"go.abhg.dev/gs/internal/git"
+	"go.abhg.dev/gs/internal/handler/delete"
 	"go.abhg.dev/gs/internal/must"
 	"go.abhg.dev/gs/internal/silog"
 	"go.abhg.dev/gs/internal/spice"
@@ -36,10 +37,10 @@ func (cmd *stackDeleteCmd) Run(
 	ctx context.Context,
 	log *silog.Logger,
 	view ui.View,
-	repo *git.Repository,
 	wt *git.Worktree,
 	store *state.Store,
 	svc *spice.Service,
+	handler DeleteHandler,
 ) error {
 	currentBranch, err := wt.CurrentBranch(ctx)
 	if err != nil {
@@ -79,8 +80,8 @@ func (cmd *stackDeleteCmd) Run(
 		return errors.New("use --force to confirm deletion")
 	}
 
-	return (&branchDeleteCmd{
+	return handler.DeleteBranches(ctx, &delete.Request{
 		Branches: stack,
-		Force:    true,
-	}).Run(ctx, log, view, repo, wt, store, svc)
+		Force:    cmd.Force,
+	})
 }
